arousal.df(s20x)	R Documentation

Changes in Pupil Size with Emotional Arousal

Description

Data from an experiment to measure the effect of different images on emotional arousal, by measuring changes in pupil diameter. The experiment used 20 males and 20 females. Images included a nude man, nude woman, infant, and a landscape.

Usage

data (arousal.df)

Format

A data frame with 160 observations on 3 variables.

[,1] 	 arousal 	 numeric 	 Change in the subject's pupil size

[,2] 	 gender 	 factor 	 Subject's gender (female, male)

[,3] 	 picture 	 factor 	 Picture shown to subject (infant, landscape, nude female, nude male)

[Package s20x version 3.1-6 Index]