autocor.plot(s20x)	R Documentation

Autocorrelation Plot

Description

Plots current vs lagged residuals along with quadrants dividing these residuals about the value zero.

Usage

autocor.plot(fit)

Arguments

fit 	 output from the function "lm()".

Value

Plots current vs lagged residuals along with quadrants dividing these residuals about the value zero.

Examples

data(airpass.df)
time<-1:144
airpass.fit<-lm(passengers~time, data = airpass.df)
autocor.plot(airpass.fit)

[Package s20x version 3.1-6 Index]