body.df(s20x)	R Documentation

Body Image and Ethnicity

Description

Data collected to examine how women from various ethnic groups rate their body image. All subjects were slightly underweight for their body size.

Usage

data (body.df)

Format

A data frame with 246 observations on 8 variables.

[,1] 	 ethnicity 	 factor 	 Subject's ethnicity (Asian, Europn, Maori, Pacific)

[,2] 	 married 	 . 	 .

[,3] 	 bodyim 	 factor 	 Subject's rating of themself (slight.uw, right, slight.ow, mod.ow, very.ow)

[,4] 	 sm.ever 	 . 	 .

[,5] 	 weight 	 . 	 .

[,6] 	 height 	 . 	 .

[,7] 	 age 	 . 	 .

[,8] 	 stressgp 	 . 	 .

[Package s20x version 3.1-6 Index]