books.df(s20x)	R Documentation

Books Data

Description

This data consists of 50 sentence lengths from each of 8 books. The books “Disclosure” and “Rising Sun” were written by Michael Crichton, whilst the others “Four Past Midnight”, “The Dark Half”, “ Eye of the Dragon”, “The Shining”, “The Stand” and “The Tommy-Knockers” where written by Stephen King. The pages and sentences where chosen using a multistage design where the pages where selected at random, and then sentences within each page were selected at random. These data were collected by James Curran.

Usage

data (books.df)

Format

The data frame consists of 400 observations on 2 variables.

[,1] 	 length 	 integer

[,2] 	 book 	 factor 	 levels (4.Past.Mid, Dark.Half, Disclosure, Eye.Drag,

Rising.Sun, Shining, Stand, T.Knock)

[Package s20x version 3.1-6 Index]