bursary.df(s20x)	R Documentation

Bursary Results for Auckland Secondary Schools

Description

Data for the 2001 Bursary results for 75 secondary schools in the Auckland area. For each school the decile rating of the school is recorded along with the percentage of eligible students who gain a B Bursary or better.

Usage

data (bursary.df)

Format

A data frame with 75 observations on 2 variables.

[,1] 	 decile 	 numeric 	 Decile rating of the school

[,2] 	 pass.rate 	 numeric 	 Percentage of eligible students who gained a "B" Busary or better

[Package s20x version 3.1-6 Index]