butterfat.df(s20x)	R Documentation

Butterfat Data

Description

This data gives the mean percentage of butterfat produced by different Canadian pure-bred diary cattle. There are five different breeds and two age groups, two years old and greater than five years old. For each combination of breed and age, there are measurements for 10 cows.

Usage

data (butterfat.df)

Format

A data frame with 100 observations on 3 variables.

[,1] 	 Butterfat 	 numeric 	 Mean percentage of butterfat per cow

[,2] 	 Breed 	 factor 	 Breed (ayrshire, canadian, guernesy, holst.fres, jersey)

[,3] 	 Age 	 factor 	 Age group (2yo, mature)

Source

A Handbook of Small Data Sets

References

Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. & Ostrowski, E. (1994). A Handbook of Small Data Sets. Boca Raton, Florida: Chapman and Hall/CRC.

Sokal, R.R. & Rohlf, F.J. (1981). Biometry, 2nd edition. San Francisco: W.H. Freeman, 368.

[Package s20x version 3.1-6 Index]