camplake.df(s20x)	R Documentation

Age and Length of Camp Lake Bluegills

Description

66 bluegills were captured from Camp Lake, Minnesota. For each bluegill we have the length of the fish, its age in years and its age in scale radius.

Usage

data (camplake.df)

Format

A data frame with 66 observations on 3 variables.

[,1] 	 Age 	 numeric 	 Age of fish (years)

[,2] 	 Scale.Radius 	 numeric 	 Age of fish (radius of the key scale (mm/100) )

[,3] 	 Length 	 numeric 	 Length at capture (mm)

[Package s20x version 3.1-6 Index]