ciReg(s20x)	R Documentation

Confidence Intervals for Regression models

Description

Calculates and prints the confidence intervals for the fitted model.

Usage

ciReg(fit, conf.level=0.95, print.out=TRUE)

Arguments

fit 	 an lm object, i.e. the output from "lm()".

conf.level 	 confidence level of the intervals.

print.out 	 if TRUE, print out the output on the screen.

Value

The function returns a two-column matrix containing the upper and lower endpoints of the intervals.

See Also

"lm", "summary", "anova".

Examples

##Peruvian Indians data
data(peru.df)
fit<-lm(BP ~ age + years + weight + height, data = peru.df)
ciReg(fit)

[Package s20x version 3.1-6 Index]