computer.df(s20x)	R Documentation

Computer Questionnaire

Description

Data from a test to see if a questionnaire was properly designed. The questionnaire measures managers' technical knowledge of computers. The test has 19 managers complete the questionnaire as well as rate their own technical expertise.

Usage

data (computer.df)

Format

A data frame with 19 observations on 2 variables.

[,1] 	 score 	 numeric 	 Questionnaire score

[,2] 	 selfassess 	 ordered factor 	 Self-assessed level of expertise (1 = low, 2 = medium, 3 = high)

[Package s20x version 3.1-6 Index]