course.df(s20x)	R Documentation

Stats 20x Summer School Data

Description

Data from a summer school Stats 20x course. Each observation represents a single student.

Usage

data (course.df)

Format

A data frame with 146 observations on 15 variables.

[,1] 	 Grade 	 factor 	 Final grade for the course (A, B, C, D)

[,2] 	 Pass 	 factor 	 Passed the course (No, Yes)

[,3] 	 Exam 	 numeric 	 Mark in the final exam

[,4] 	 Degree 	 factor 	 Degree enrolled in (BA, BCom, BSc, Other)

[,5] 	 Gender 	 factor 	 Gender (Female, Male)

[,6] 	 Attend 	 factor 	 Regularly attended class (No, Yes)

[,7] 	 Assign 	 numeric 	 Assignment mark

[,8] 	 Test 	 numeric 	 Test mark

[,9] 	 B 	 numeric 	 Mark for the short answer section of the exam

[,10] 	 C 	 numeric 	 Mark for the long answer section of the exam

[,11] 	 MC 	 numeric 	 Mark for the multiple choice section of the exam

[,12] 	 Colour 	 factor 	 Colour of the exam booklet (Blue, Green, Pink, Yellow)

[,13] 	 Stage1 	 factor 	 Stage one grade (A, B, C)

[,14] 	 Years.Since 	 numeric 	 Number of years since doing Stage 1

[,15] 	 Repeat 	 factor 	 Repeating the paper (No, Yes)

[Package s20x version 3.1-6 Index]