course2way.df(s20x)	R Documentation

Exam Mark Gender and Attendance for Stats 20x Summer School Students

Description

Data from a summer school Stats 20x course. Each observation represents a single student. It is of interest to see if there is a relationship between a student's final examination mark and both their gender and whether they regularly attend lectures.

Usage

data (course2way.df)

Format

A data frame with 40 observations on 3 variables.

[,1] 	 Exam 	 numeric 	 Final exam mark (out of 100)

[,2] 	 Gender 	 factor 	 Gender (Female, Male)

[,2] 	 Attend 	 factor 	 Regularly attended or not (No, Yes)

[Package s20x version 3.1-6 Index]