fire.df(s20x)	R Documentation

Fire Damage and Distance from the Fire Station

Description

House damage and distance from the fire station, of 15 house fires. Data collected by an insurance company for homes in a particular area.

Usage

data (fire.df)

Format

A data frame with 15 observations on 2 variables.

[,1] 	 damage 	 numeric 	 Damage ($000s)

[,2] 	 distance 	 numeric 	 Distance from the fire station (miles)

[Package s20x version 3.1-6 Index]