fruitfly.df(s20x)	R Documentation

Fruitfly Data

Description

This data gives fecundity for female fruitflies, Drosophila melanogaster. The fecundity is the number of eggs laid, per day, for the fruitfly's first 14 days of life. There are three strains: A control group, NS, Nonselected Strain, as well as RS, a strain bred for resistance to DDT and SS, a strain bred for susceptibility to DDT. Each strain contains 25 measurements. It is of interest to compare the level of fecundity across strains.

Usage

data (fruitfly.df)

Format

A data frame with 75 observations on 2 variables.

[,1] 	 fecundity 	 numeric 	 Number of eggs laid, per day, per fruitfly

[,2] 	 strain 	 factor 	 Strain of fruitfly (NS, RS, SS)

Source

A Handbook of Small Data Sets

References

Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. & Ostrowski, E. (1994). A Handbook of Small Data Sets. Boca Raton, Florida: Chapman and Hall/CRC.

Sokal, R.R. & Rohlf, F.J. (1981). Biometry, 2nd edition. San Francisco: W.H. Freeman, 239.

[Package s20x version 3.1-6 Index]