levene.test(s20x)	R Documentation

Levene test for the ANOVA Assumption

Description

Perform a Levene test for equal group variances in both one-way and two-way ANOVA. A table with the results is (normally) displayed.

Usage

levene.test(formula, data, digit = 5, show.table = TRUE)

Arguments

formula 	 a symbolic description of the model to be fitted: response ~ fac1 + fac2.

data 	 an optional data frame containing the variables in the model.

digit 	 the number of decimal places to display.

show.table 	 If this argument is FALSE then the output will be suppressed

Value

df 	 degrees of freedom.

ss 	 sum squares.

ms 	 mean squares.

f.value 	 F-statistic value.

p.value 	 P-value.

See Also

"crossFactors", "anova".

Examples

##
data(computer.df)
levene.test(score ~ factor(selfassess), computer.df)

[Package s20x version 3.1-6 Index]