nail.df(s20x)	R Documentation

Nail Polish Data

Description

These data were collected to determine whether quick drying nail polish or regular nail polish dried faster. The time for each type of nail polish to dry was recorded.

Usage

data (nail.df)

Format

A data frame with 60 observations on 2 variables.

[,1] 	 polish 	 factor 	 Type of polish (Regular, Quick)

[,2] 	 dry 	 integer 	 Time (in seconds) for the polish to dry

[Package s20x version 3.1-6 Index]