onewayPlot(s20x)	R Documentation

One-way Analysis of Variance Plot

Description

Displays stripplot/boxplot of the reponse variable with intervals by factor levels. It is used as part of a one-way anova analysis.

Usage

onewayPlot(x,...)
## S3 method for class 'formula':
onewayPlot(formula, data = parent.frame(),
                             conf.level = 0.95,
                             interval.type = "tukey",
                             pooled = TRUE, strip = TRUE, vert = TRUE,
                             verbose = FALSE,
                             ylabel = deparse(terms(formula)[[2]]),
                             flabel = deparse(terms(formula)[[3]]),...)
## S3 method for class 'lm':
onewayPlot(x,...)

Arguments

formula 	 a symbolic description of the model to be fit.

x 	 either a formula or a lm object

conf.level 	 confidence level of the intervals.

interval.type 	 three options for intervals appearing on plot: "hsd","lsd" or "ci".

pooled 	 two options: pooled or unpooled standard deviation used for plotted intervals.

strip 	 if strip=F, boxplots are displayed instead.

vert 	 if vert=F, horizontal stripplots are displayed instead (boxplots can only be displayed vertically).

verbose 	 if true, print intervals on console.

ylabel 	 can be used to replace variable name of y by another string.

flabel 	 can be used to replace variable name of f by another string.

data 	 an optional data frame in which to evaluate the formula.

... 	 optional arguments.

See Also

"summary1way", "welch.test".

Examples

##see example in "summary1way"

##sheep data:
data(computer.df)
onewayPlot(score~selfassess, data = computer.df)

[Package s20x version 3.1-6 Index]