peru.df(s20x)	R Documentation

Peruvian Indians

Description

A random sample of Peruvian Indians born in the Andes mountains, but who have since migrated to lower altitudes. The sample was collected to assess the long term effects of altitude on blood pressure.

Usage

data(peru.df)

Format

A data frame with 39 observations on 5 variables.

[,1] 	 age 	 numeric 	 Subject's age

[,2] 	 years 	 numeric 	 Number of years since migration

[,3] 	 weight 	 numeric 	 Subject's weight (kg)

[,4] 	 height 	 numeric 	 Subject's height (mm)

[,5] 	 BP 	 numeric 	 Subject's systolic blood pressure (mm Hg)

[Package s20x version 3.1-6 Index]