rain.df(s20x)	R Documentation

Cloud Seeding and Levels of Rainfall

Description

Data from an experiment to see if seeding clouds with Silver Nitrate effects the amount of rainfall.

Usage

data (rain.df)

Format

A data frame with 50 observations on 2 variables.

[,1] 	 rain 	 numeric 	 Amount of rain

[,2] 	 seed 	 factor 	 Whether the clouds are seeded or not (seeded, unseeded)

Source

Chambers, Cleveland, Kleiner, Tukey. (1983). Graphical Methods for Data Analysis.

[Package s20x version 3.1-6 Index]