skulls.df(s20x)	R Documentation

Skulls Data

Description

Male Egyptian skulls from five different epochs. Each skull has had four measurements taken of it, BH, Basibregmatic Height, BL, Basialveolar Length, MB, Maximum Breadth and NH, Nasal Height. It is of interest to investigate the change in shape over time. A gradual change, would indicate inbreeding of the populations. This data only includes the maximum breadth measurements.

Usage

data (skulls.df)

Format

A data frame with 150 observations on 2 variables.

[,1] 	 measurement 	 integer

[,2] 	 year 	 integer

Source

A Handbook of Small Data Sets

References

Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. & Ostrowski, E. (1994). A Handbook of Small Data Sets. Boca Raton, Florida: Chapman and Hall/CRC.

Thomson, A. & Randall-Maciver, R. (1905). Ancient Races of the Thebaid. Oxford: Oxford University Press.

[Package s20x version 3.1-6 Index]