soyabean.df(s20x)	R Documentation

Soya Bean Yields

Description

Data from an experiment to examine the effects of different planting times on the yield of soya beans, given four different cultivars.

Usage

data(soyabean.df)

Format

A data frame with 32 observations on 3 variables.

[,1] 	 yield 	 numeric 	 Yield of each plant

[,2] 	 cultivar 	 factor 	 Cultivar used (cult1, cult2, cult3, cult4)

[,3] 	 planttime 	 factor 	 Month of planting (Novemb, Decemb)

Source

Littler, R. University of Waikato

[Package s20x version 3.1-6 Index]