toothpaste.df(s20x)	R Documentation

Crest Toothpaste

Description

Two random samples of households, one of households who purchase Crest toothpaste and one of households who do not. For each household the age is recorded of the person responsible for purchasing the toothpaste.

Usage

data (toothpaste.df)

Format

A data frame with 20 observations on 2 variables.

[,1] 	 purchasers 	 numeric 	 Age of the person in the household responsible for purchases of Crest

[,2] 	 nonpurchasers 	 numeric 	 Age of the person in the household responsible for purchases of other brands of toothpaste

[Package s20x version 3.1-6 Index]