trendscatter(s20x)	R Documentation

Trend and scatter plot

Description

Plots a scatter plot for the variables x, y along with a lowess smooth for the underlying trend. One standard deviation error bounds for the scatter about this trend are also plotted.

Usage

trendscatter(x, ...)
## Default S3 method:
trendscatter (x, y = NULL, f = 0.5, xlab = NULL, ylab = NULL, ...)
## S3 method for class 'formula':
trendscatter (x, f = 0.5, data, xlab = NULL, ylab = NULL, ...)

Arguments

x 	 the coordinates of the points in the scatter plot. Alternatively, a function.

y 	 the y coordinates of the points in the plot, ignored if 'x' is a function.

f 	 the smoother span. This gives the proportion of points in the plot which influence the smooth at each value. Larger values give more smoothness.

data 	 an optional data frame containing the variables in the model.

xlab 	 a title for the x axis: see title.

ylab 	 a title for the y axis: see title.

... 	 Optional arguments

Value

Returns the plot.

See Also

"residPlot"

Examples

# A simple polynomial
x<-rnorm(100)
e<-rnorm(100)
y<-2+3*x-2*x^2+4*x^3+e
trendscatter(y~x)

# An exponential growth curve
e<-rnorm(100,0,0.1)
y<-exp(5+3*x+e)
trendscatter(log(y)~x)

# Peruvian Indians data
data(peru.df)
trendscatter(BP~weight, data=peru.df)

# Note: this usage is deprecated
with(peru.df,trendscatter(weight,BP))

[Package s20x version 3.1-6 Index]