version(s20x)	R Documentation

s20x pacakge version number

Description

Returns the version number of the s20x package. This is useful if a student is has problems runnning commands and the maintainer needs to check the version number.

Usage

version()

Examples

version()

[Package s20x version 3.1-4 Index]