zoo.df(s20x)	R Documentation

Zoo Attendance during an Advertising Campaign

Description

Data for 455 days of attendance records for Auckland Zoo, from January 1, 1993. Note that only 440 values are given due to missing values. It was of interest to assess whether an advertising campaign was effective in increasing attendance.

Usage

data (zoo.df)

Format

A data frame with 440 observations on 6 variables.

[,1] 	 attendance 	 numeric 	 Number of visitors

[,2] 	 time 	 numeric 	 Time in days since the start of the study

[,3] 	 sun.yesterday 	 numeric 	 Hours of sunshine the previous day

[,4] 	 tv.adds 	 numeric 	 Average spending on TV advertising in the previous week ($000 per day)

[,5] 	 nice.day 	 factor 	 Assessment based on number of hours of sunshine (0 = No, 1 = Yes)

[,6] 	 day.type 	 factor 	 Type of day (1 = ordinary weekday, 2 = weekend day, 3 = school holiday weekday, 4 = public holday)

[Package s20x version 3.1-6 Index]