boxqq(s20x)	R Documentation

Box plots and normal quantile-quantile plots

Description

Draws boxplots and normal quantile quantile plots of x for each value of the grouping variable g

Usage

boxqq(formula, ...)
## S3 method for class 'formula':
boxqq(formula, data = NULL, ...)

Arguments

formula 	 A symbolic specification of the form `x ~ g' can be given, indicating the observations in the vector `x' are to be grouped according to the levels of the factor `g'. `NA's are allowed in the data.

data 	 An optional data frame in which to evaluate the formula.

... 	 Arguments to be passed to methods, such as graphical parameters (see par).

Value

Returns the plot.

Examples

## Zoo data
data(zoo.df)
boxqq(attendance~day.type, data = zoo.df)

[Package s20x version 3.1-6 Index]