chalk.df(s20x)	R Documentation

Chalk Data

Description

These data involve 11 laboratories and 2 brands of chalk. The laboratories tested the density of the chalk. The main interest was whether the different laboratories yielded the same density for the two different types of chalk.

Usage

data (chalk.df)

Format

A data frame with 66 observations on 3 variables.

[,1] 	 Density 	 numeric 	 Density of the chalk

[,2] 	 Lab 	 integer 	 Laboratory where testing done

[,3] 	 Chalk 	 factor 	 Chalk tested (A, B)

[Package s20x version 3.1-6 Index]