mergers.df(s20x)	R Documentation

Merger Days

Description

A random selection of 38 consummated mergers from the USA, 1982, giving the number of days between the date the merger was announced and the date the merger became effective.

Usage

data (mergers.df)

[Package s20x version 3.1-6 Index]