multipleComp(s20x)	R Documentation

Multiple Comparisons

Description

Calculates and prints the estimate, multiple 95% confidence intervals; unadjusted, Tukey and Bonferroni p-values for all possible differences in means in a one-way ANOVA.

Usage

multipleComp(fit,conf.level=0.95)

Arguments

fit 	 output from the command "lm()".

conf.level 	 confidence level for the confidence interval, expressed as a percentage.

Value

Returns a list of estimates, confidence intervals and p-values.

Examples

## computer data
data(computer.df)
fit <- lm(score ~ factor(selfassess), data = computer.df)
multipleComp(fit)

[Package s20x version 3.1-6 Index]