pairs20x(s20x)	R Documentation

Pairwise Scatter Plots with Histograms and Correlations

Description

Plots pairwise scatter plots with histograms and correlations for the data frame.

Usage

pairs20x(x, ...)

Arguments

x 	 a data frame.

... 	 optional argumments which are passed to the generic pairs function.

Value

Returns the plots.

See Also

"pairs", "panel.smooth", "panel.cor", "panel.hist"

Examples

##peruvian indians
data(peru.df)
pairs20x(peru.df)

[Package s20x version 3.1-6 Index]