seeds.df(s20x)	R Documentation

Seeds Data

Description

These data record the number of seeds (out of 100) that germinated when given different amounts of water. The seeds were either exposed to light or kept in the dark. Four identical boxes were used for each combination of water and light

Usage

data (seeds.df)

Format

A data frame with 48 observations on 3 variables.

[,1] 	 Light 	 integer 	 Seeds exposed to light (N=No, Y=Yes)

[,2] 	 Water 	 integer 	 Amount of water, higher levels correspond to more water (1, 2, 3, 4, 5, 6)

[,3] 	 Count 	 integer 	 Number of seeds that germinated (out of 100)

[Package s20x version 3.1-6 Index]