skewness(s20x)	R Documentation

Skewness Statistic

Description

Calculates the skewness statistic of the data in "x". Values close to zero correspond to reasonably symmetric data, positive values of this measure indicate right-skewed data whereas negative values indicate left-skewness.

Usage

skewness(x)

Arguments

x 	 vector containing the data.

Value

Returns the value of the skewness.

Examples

##Merger data:
data(mergers.df)
skewness(mergers.df$mergerdays)

[Package s20x version 3.1-6 Index]