summary1way(s20x)	R Documentation

One-way Analysis of Variance Summary

Description

Displays summary information for a one-way anova analysis. The lm object must come from a numerical response variable and a single factor. The output includes: (i) anova table; (ii) numeric summary; (iii) table of effects; (iv) plot of data with intervals.

Usage

summary1way(fit, digit = 5, conf.level = 0.95, inttype = "tukey",
             pooled = TRUE, print.out = TRUE, draw.plot = TRUE, ...)

Arguments

fit 	 an lm object, i.e. the output from "lm()".

digit 	 decimal numbers after the point.

conf.level 	 confidence level of the intervals.

inttype 	 three options for intervals appeared on plot: "hsd","lsd" or "ci".

pooled 	 two options: pooled or unpooled standard deviation used for plotted intervals.

print.out 	 if T, print out the output on the screen.

draw.plot 	 if T, plot data with intervals.

... 	 more options.

Value

Df 	 degrees of freedom for regression, residual and total.

Sum of Sq 	 sum squares for regression, residual and total.

Mean Sq 	 mean squares for regression and residual.

F value 	 F-statistic value.

Pr(F)

Main Effect

Group Effects

See Also

"summary2way", "anova", "aov", "dummy.coef", "plot.oneway"

Examples

attitudes <- c(5.2,5.2,6.1,6,5.75,5.6,6.25,6.8,6.87,7.1,
               6.3,6.35,5.5,5.75,4.6,5.36,5.85,5.9)
l <- rep(c("Gp1","Gp2","Gp3"),rep(6,3))
l <- factor(l)
f <-lm(attitudes ~ l)
result <- summary1way(f)
result

[Package s20x version 3.1-6 Index]