teach.df(s20x)	R Documentation

Comparison of Three Teaching Methods

Description

Data from an experiment to assess the impact of three different teaching methods on language ability. 30 students were randomly allocated into three groups, one for each method. The students' IQ before instruction and a language test score after instruction were recorded.

Usage

data (teach.df)

Format

A data frame with 30 observations on 3 variables.

[,1] 	 lang 	 numeric 	 Language test score after instruction

[,2] 	 IQ 	 numeric 	 Student's IQ

[,3] 	 method 	 factor 	 Teaching method (1, 2, 3)

[Package s20x version 3.1-6 Index]