thyroid.df(s20x)	R Documentation

Effect of a New Drug on Thyroid Weights

Description

Data from an experiment to asses the effect of a new drug on the weight of the thyroid gland using 16 laboratory animals. The animals were randomly assigned into either a control group, or a treatment group, and each animal had its bodyweight recorded at the beginning of the experiment and its thyroid weight measured at the end of the experiment.

Usage

data(thyroid.df)

Format

A data frame with 16 observations on 3 variables.

[,1] 	 thyroid 	 numeric 	 Weight of thyroid gland after 7 days (mg)

[,2] 	 body 	 numeric 	 Animal body weight before experiment began (g)

[,3] 	 group 	 factor 	 Animal's group (1 = control, 2 = drug)

[Package s20x version 3.1-6 Index]